strip out the interruption-related agent code
It was a fun experiment, but it adds too much complexity and weird error
opportunities - and it's already trivial to just use the pause button to
"stop" the agent from typing. only thing we're missing by removing this
is having the message-in-progress persisted, but the whole message will
be persisted when it gets completed in the separate conversation slot
anyways.